make header fixed
- make header position fixed >> - only content will be scrollable
