+++ app/src/main/java/com/app/moviematrix/MainActivity.kt
@@ -1,23 +1,13 @@
 package com.app.moviematrix
 
-import android.app.Activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import androidx.activity.SystemBarStyle
 import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.BottomNavigationItem
 import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.List
@@ -32,21 +22,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
-import com.app.moviematrix.ui.theme.MovieMatrixTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-//        enableEdgeToEdge(
-//            statusBarStyle = SystemBarStyle.dark(
-//                ContextCompat.getColor(applicationContext, R.color.background)
-//            )
-//        )
-
         //statusBarColor
         window.statusBarColor = ContextCompat.getColor(applicationContext, R.color.background)
 
@@ -106,6 +87,14 @@ class MainActivity : ComponentActivity() {
             Box(
                 modifier = Modifier
                     .padding(innerPadding)
+                    .background(
+                        Color(
+                            ContextCompat.getColor(
+                                applicationContext,
+                                R.color.background
+                            )
+                        )
+                    )
             ) {
                 when (selectedItem) {
                     0 -> Home()
@@ -115,32 +104,6 @@ class MainActivity : ComponentActivity() {
                 }
             }
 
-        }
-    }
-
-    @Composable
-    fun Home() {
-        Column(
-            modifier = Modifier
-                .background(
-                    Color(
-                        ContextCompat.getColor(
-                            applicationContext,
-                            R.color.background
-                        )
-                    )
-                )
-                .fillMaxHeight()
-                .fillMaxWidth()
-                .padding(horizontal = 15.dp)
-                .verticalScroll(rememberScrollState())
-        ) {
-            Header()
-            SliderBanner()
-            TrendingPersonList()
-            TrendingMovies()
-            PopularMovies()
-            TopRatedTvShow()
         }
     }
 }
